Update to v106r175 release.
byuu says:

  - ruby: fullscreen support for Windows OpenGL 3.2, DirectDraw 7.0, GDI drivers
  - ruby: output(width, height) support for all drivers on all platforms
  - hiro: improve focus capture for Canvas and Viewport widgets
  - hiro: added two hotfixes for the macOS Cocoa target [Sintendo]
  - higan, bsnes: focus the viewport after leaving fullscreen exclusive mode
  - freebsd: moved from GCC 8.2.0 to Clang 6.0.1
  - higan: added video display emulation option to Famicom and Mega Drive cores

The reason I moved to Clang was because GCC keeps deadlocking my FreeBSD system.
I don't know if it's GCC's fault, or suddenly running 32 copies of any high-CPU
usage program, heh. But at any rate, it's worth a try. The performance is the
same, but compilation takes a tiny bit longer with Clang.

+++ higan/component/processor/wdc65816/disassembler.cpp
@@ -46,7 +46,7 @@ auto WDC65816::dreadl(uint24 addr) -> uint24 {
   return data;
 }
 
-auto WDC65816::decode(uint8 mode, uint24 addr) -> uint24 {
+auto WDC65816::decode(uint8 mode, uint24 addr) -> uint {
   uint24 a = 0;
 
   switch(mode) {
@@ -123,7 +123,7 @@ auto WDC65816::decode(uint8 mode, uint24 addr) -> uint24 {
     break;
   }
 
-  return a;
+  return (uint24)a;
 }
 
 auto WDC65816::disassemble() -> string {
@@ -136,10 +136,10 @@ auto WDC65816::disassemble(uint24 addr, bool e, bool m, bool x) -> string {
   uint24 pc = addr;
   s = {hex(pc, 6), " "};
 
-  uint8 op  = dreadb(pc); pc.bits(0,15)++;
-  uint8 op0 = dreadb(pc); pc.bits(0,15)++;
-  uint8 op1 = dreadb(pc); pc.bits(0,15)++;
-  uint8 op2 = dreadb(pc);
+  uint8_t op  = dreadb(pc); pc.bits(0,15)++;
+  uint8_t op0 = dreadb(pc); pc.bits(0,15)++;
+  uint8_t op1 = dreadb(pc); pc.bits(0,15)++;
+  uint8_t op2 = dreadb(pc);
 
   #define op8  ((op0))
   #define op16 ((op0) | (op1 << 8))
