+++ higan/component/processor/wdc65816/wdc65816.hpp
@@ -208,7 +208,7 @@ struct WDC65816 {
   auto dreadb(uint24 addr) -> uint8;
   auto dreadw(uint24 addr) -> uint16;
   auto dreadl(uint24 addr) -> uint24;
-  auto decode(uint8 mode, uint24 addr) -> uint24;
+  auto decode(uint8 mode, uint24 addr) -> uint;
 
   struct Flags {
     bool c;  //carry
